+++ detox/src/Detox.js
@@ -3,10 +3,10 @@ const util = require('util');
 const logger = require('./utils/logger');
 const log = require('./utils/logger').child({ __filename });
 const Device = require('./devices/Device');
-const IosDriver = require('./devices/IosDriver');
-const SimulatorDriver = require('./devices/SimulatorDriver');
-const EmulatorDriver = require('./devices/EmulatorDriver');
-const AttachedAndroidDriver = require('./devices/AttachedAndroidDriver');
+const IosDriver = require('./devices/drivers/IosDriver');
+const SimulatorDriver = require('./devices/drivers/SimulatorDriver');
+const EmulatorDriver = require('./devices/drivers/EmulatorDriver');
+const AttachedAndroidDriver = require('./devices/drivers/AttachedAndroidDriver');
 const DetoxRuntimeError = require('./errors/DetoxRuntimeError');
 const argparse = require('./utils/argparse');
 const configuration = require('./configuration');
