+++ detox/src/devices/drivers/EmulatorDriver.js
@@ -1,11 +1,11 @@
 const _ = require('lodash');
 const path = require('path');
-const Emulator = require('./android/Emulator');
-const EmulatorTelnet = require('./android/EmulatorTelnet');
-const DetoxRuntimeError = require('../errors/DetoxRuntimeError');
-const Environment = require('../utils/environment');
-const retry = require('../utils/retry');
-const sleep = require('../utils/sleep');
+const Emulator = require('../android/Emulator');
+const EmulatorTelnet = require('../android/EmulatorTelnet');
+const DetoxRuntimeError = require('../../errors/DetoxRuntimeError');
+const Environment = require('../../utils/environment');
+const retry = require('../../utils/retry');
+const sleep = require('../../utils/sleep');
 const AndroidDriver = require('./AndroidDriver');
 const ini = require('ini');
 const fs = require('fs');
