+++ detox/src/devices/drivers/IosDriver.js
@@ -1,18 +1,18 @@
 const path = require('path');
 const fs = require('fs');
-const log = require('../utils/logger').child({ __filename });
+const log = require('../../utils/logger').child({ __filename });
 const DeviceDriverBase = require('./DeviceDriverBase');
-const InvocationManager = require('../invoke').InvocationManager;
-const invoke = require('../invoke');
-const GREYConfigurationApi = require('./../ios/earlgreyapi/GREYConfiguration');
-const GREYConfigurationDetox = require('./../ios/earlgreyapi/GREYConfigurationDetox');
-const EarlyGrey = require('./../ios/earlgreyapi/EarlGrey');
+const InvocationManager = require('../../invoke').InvocationManager;
+const invoke = require('../../invoke');
+const GREYConfigurationApi = require('../../ios/earlgreyapi/GREYConfiguration');
+const GREYConfigurationDetox = require('../../ios/earlgreyapi/GREYConfigurationDetox');
+const EarlyGrey = require('../../ios/earlgreyapi/EarlGrey');
 
 class IosDriver extends DeviceDriverBase {
   constructor({ client, emitter }) {
     super({ client, emitter });
 
-    this.expect = require('../ios/expect');
+    this.expect = require('../../ios/expect');
     this.expect.setInvocationManager(new InvocationManager(client));
   }
