+++ detox/src/devices/drivers/SimulatorDriver.js
@@ -3,14 +3,14 @@ const path = require('path');
 const fs = require('fs');
 const _ = require('lodash');
 const IosDriver = require('./IosDriver');
-const AppleSimUtils = require('./AppleSimUtils');
-const configuration = require('../configuration');
-const environment = require('../utils/environment');
-const DeviceRegistry = require('./DeviceRegistry');
+const AppleSimUtils = require('../ios/AppleSimUtils');
+const configuration = require('../../configuration');
+const environment = require('../../utils/environment');
+const DeviceRegistry = require('../DeviceRegistry');
 
-const SimulatorLogPlugin = require('../artifacts/log/ios/SimulatorLogPlugin');
-const SimulatorScreenshotPlugin = require('../artifacts/screenshot/SimulatorScreenshotPlugin');
-const SimulatorRecordVideoPlugin = require('../artifacts/video/SimulatorRecordVideoPlugin');
+const SimulatorLogPlugin = require('../../artifacts/log/ios/SimulatorLogPlugin');
+const SimulatorScreenshotPlugin = require('../../artifacts/screenshot/SimulatorScreenshotPlugin');
+const SimulatorRecordVideoPlugin = require('../../artifacts/video/SimulatorRecordVideoPlugin');
 
 class SimulatorDriver extends IosDriver {
