+++ detox/src/devices/ios/AppleSimUtils.js
@@ -2,9 +2,9 @@ const process = require('process');
 const _ = require('lodash');
 const tempfile = require('tempfile');
 const cpp = require('child-process-promise');
-const exec = require('../utils/exec');
-const retry = require('../utils/retry');
-const environment = require('../utils/environment');
+const exec = require('../../utils/exec');
+const retry = require('../../utils/retry');
+const environment = require('../../utils/environment');
 
 class AppleSimUtils {
   async setPermissions(udid, bundleId, permissionsObj) {
