+++ detox/src/devices/ios/xcrunSimctlList.mock.json
